feat: split to handler fns and move annotated collections to its own file
