+++ types/types.ts
@@ -177,7 +177,7 @@ export interface AnnotatedCollection {
   id: string;
   name: string;
   description?: string;
-  collection_type: "incident" | "wildlife_sighting" | "research_plot";
+  collection_type: "incident";
   status: "active" | "archived" | "draft";
   created_by: string;
   created_at: string;
@@ -196,21 +196,6 @@ export interface Incident {
     documents?: string[];
     testimonies?: string[];
   };
-}
-
-export interface WildlifeSighting {
-  collection_id: string;
-  species?: string;
-  observation_type?:
-    | "camera_trap"
-    | "direct_observation"
-    | "track"
-    | "scat"
-    | "call";
-  behavior_notes?: string;
-  habitat_type?: string;
-  weather_conditions?: string;
-  observer_confidence?: "high" | "medium" | "low";
 }
 
 export interface CollectionEntry {
